Working on getting docker running on azure

+++ a2/vm.py
@@ -21,8 +21,13 @@ from azure.mgmt.compute import ComputeManagementClient
 from azure.mgmt.compute.models import DiskCreateOption
 from msrestazure.azure_exceptions import CloudError
 from pexpect import pxssh
+from multiprocessing import Process
+
 
 DISPLAY_STORAGE = True
+AWS_KEY = 'keys/testing6.pem'
+AZURE_PUB_KEY = 'keys/azureKey.pub'
+AZURE_KEY = 'keys/azureKey'
 
 awsos = {}
 awsos['ami-0a887e401f7654935'] = 'Amazon Linux'
@@ -32,13 +37,14 @@ awsos['ami-0df6cfabfbe4385b7'] = 'SUSE Linux'
 awsos['ami-07ebfd5b3428b6f4d'] = 'Ubuntu Server'
 
 #region [rgba(100,202,240,0.1)] Docker
-def buildDockerScript(id, os, docker):
+def buildDockerScript(id, os, docker, flag):
     cmd = 'apt'
     if os == 'Amazon Linux':
         cmd = 'yum'
     script = "#!/bin/bash\n"
     script = script + 'sudo '+cmd+' update -y \n'
-    script = script + 'sudo '+cmd+' update -y \n'
+    if (flag):
+        script = script + 'curl -sSL https://get.docker.com/ | sh \n'
     script = script + 'sudo '+cmd+' install -y docker \n'
     script = script + 'sudo service docker start \n'
     for image in docker:
@@ -98,7 +104,7 @@ def createAWSVM(win, vm, docker):
     #Create instance 
     win.addstr('Creating VM instance\n')
     win.refresh()
-    script = buildDockerScript(vm[2], awsos[vm[1]], docker)
+    script = buildDockerScript(vm[2], awsos[vm[1]], docker, False)
     response = ec2.create_instances(BlockDeviceMappings=[
         {
             'DeviceName': '/dev/xvda',
@@ -119,7 +125,8 @@ def killAWS():
     ids = []
     for s in instances:
         ids.append(s['InstanceId'])
-    ec2_c.terminate_instances(InstanceIds=ids)
+    if len(ids) > 0:
+        ec2_c.terminate_instances(InstanceIds=ids)
 
 def rebootAWS():
     instances = ec2.meta.client.describe_instance_status(IncludeAllInstances=True)['InstanceStatuses']
@@ -164,6 +171,8 @@ images = {}
 images['debian'] = debian
 images['redHat'] = redHat
 images['ubuntu'] = ubuntu
+
+dockerDeployments = []
 
 #https://docs.microsoft.com/en-us/python/api/azure-mgmt-network/azure.mgmt.network.v2019_11_01.operations.networkinterfacesoperations?view=azure-python
 def create_nic(resourceGroup, uid):
@@ -221,7 +230,7 @@ def create_nic(resourceGroup, uid):
     return async_nic_creation.result()
 
 def create_vm_parameters(nic_id, vm):
-    keyFile = open('keys/azureKey.pub','r')
+    keyFile = open(AZURE_PUB_KEY,'r')
     return {
         'location': 'eastus',
         'os_profile': {
@@ -254,6 +263,7 @@ def create_vm_parameters(nic_id, vm):
 def createAzureVM(win, vm, docker):
     win.addstr('\n\nCreating Azure VM: '+vm[2]+'\n')
     win.refresh()
+
     # Create a NIC
     resourceGroup = 'cis4010A2'
     id = ''
@@ -266,6 +276,7 @@ def createAzureVM(win, vm, docker):
         nic = network_client.network_interfaces.get('cis4010A2', vm[2]+'-nic')
         id = nic.id
         pass
+
     # Create Linux VM
     win.addstr('Creating machine\n')
     win.refresh()
@@ -273,53 +284,67 @@ def createAzureVM(win, vm, docker):
     async_vm_creation = compute_client.virtual_machines.create_or_update('cis4010A2', vm[2], vm_parameters)
     #async_vm_creation.wait()
 
-    # Create managed data disk
-    win.addstr('Creating disk\n')
-    win.refresh()
-    async_disk_creation = compute_client.disks.create_or_update(
-        resourceGroup,
-        vm[2]+'-disk',
-        {
-            'location': 'eastus',
-            'disk_size_gb': int(vm[6]),
-            'creation_data': {
-                'create_option': DiskCreateOption.empty
+    if vm[4] == 'Y':
+        # Create managed data disk
+        win.addstr('Creating disk\n')
+        win.refresh()
+        async_disk_creation = compute_client.disks.create_or_update(
+            resourceGroup,
+            vm[2]+'-disk',
+            {
+                'location': 'eastus',
+                'disk_size_gb': int(vm[6]),
+                'creation_data': {
+                    'create_option': DiskCreateOption.empty
+                }
             }
-        }
-    )
-    data_disk = async_disk_creation.result()
+        )
+        data_disk = async_disk_creation.result()
 
-    # Get the virtual machine by name
-    #print('\nGet Virtual Machine by Name')
-    virtual_machine = compute_client.virtual_machines.get(
-        resourceGroup,
-        vm[2]
-    )
+        # Get the virtual machine by name
+        virtual_machine = compute_client.virtual_machines.get(
+            resourceGroup,
+            vm[2]
+        )
 
-    # Attach data disk
-    win.addstr('Attaching disk\n')
-    win.refresh()
-    virtual_machine.storage_profile.data_disks.append({
-        'lun': 12,
-        'name': vm[2]+'-disk',
-        'create_option': DiskCreateOption.attach,
-        'managed_disk': {
-            'id': data_disk.id
-        }
-    })
-    async_disk_attach = compute_client.virtual_machines.create_or_update(
-        resourceGroup,
-        virtual_machine.name,
-        virtual_machine
-    )
+        # Attach data disk
+        win.addstr('Attaching disk\n')
+        win.refresh()
+        virtual_machine.storage_profile.data_disks.append({
+            'lun': 12,
+            'name': vm[2]+'-disk',
+            'create_option': DiskCreateOption.attach,
+            'managed_disk': {
+                'id': data_disk.id
+            }
+        })
+        async_disk_attach = compute_client.virtual_machines.create_or_update(
+            resourceGroup,
+            virtual_machine.name,
+            virtual_machine
+        )
     #async_disk_attach.wait()
+    dockerDeployments.append(vm[2])
 
-    
+def installDockerAzure(id, ip, key):
+    s = pxssh.pxssh()
+    vms, docker = readFiles()
+    if not s.login (server=ip, username='adminL0gin', ssh_key=key):
+        return "SSH session failed on login."+str(s)
+    else:
+        script = buildDockerScript(id, 'azure', docker, True)
+        s.sendline("echo '"+script+"' > script.sh")
+        s.prompt()
+        s.sendline('sudo chmod +x script.sh')
+        s.prompt()
+        s.sendline('sudo nohup ./script.sh &')
+        s.prompt()
+        s.logout()
+        return 'success'
+
 def killAzure():
     for vm in compute_client.virtual_machines.list_all():
         compute_client.virtual_machines.delete('cis4010A2', vm.name)
-        #TODO need to detach vm from nic before deleting
-        #network_client.virtual_networks.delete('cis4010A2', vm.name+'-nic')
 
 def rebootAzure():
     for vm in compute_client.virtual_machines.list_all():
@@ -373,6 +398,7 @@ def drawHeader(win):
 def drawFooter(win, msg):
     win.addstr('\n'+msg+'\n')
     win.addstr('\nc create VMs\n')
+    win.addstr('d to view docker\n')
     win.addstr('r to reboot\n')
     win.addstr('t to terminate\n')
     win.addstr('q to quit\n')
@@ -416,21 +442,43 @@ def drawUpdateAWS(win):
             win.addstr('\n')
 
 def drawUpdateAzure(win):
+    procs = []
     for vm in compute_client.virtual_machines.list_all():
         states = compute_client.virtual_machines.instance_view('cis4010A2', vm.name, expand='instanceView').statuses
         state = states[0].display_status
         if (len(states) > 1):
             state = states[1].display_status
+            if vm.name in dockerDeployments:
+                if 'running' in state:
+                    try:
+                        win.addstr('Installing docker images on: '+vm.name+'\n')
+                        public_ip = network_client.public_ip_addresses.get('cis4010A2', vm.name+'-pub-ip').ip_address
+                        p =  Process(target=installDockerAzure, args=(vm.name, public_ip, AZURE_KEY))
+                        p.start()
+                        procs.append(p)
+                        dockerDeployments.remove(vm.name)
+                    except Exception as  e:
+                        win.addstr('Failed to install docker images on: '+vm.name+'\n'+e+'\n')
         y, x = win.getyx()
         win.addstr(y, x, vm.name)
-        win.addstr(y, x+24, state+'\n')
+        win.addstr(y, x+24, state)
         try:
             public_ip = network_client.public_ip_addresses.get('cis4010A2', vm.name+'-pub-ip')
-            win.addstr(y, x+48, str(public_ip.ip_address))
+            win.addstr(y, x+48, str(public_ip.ip_address)+'\n')
         except:
+            win.addstr('\n')
             pass
+        y, x = win.getyx()
+        win.addstr('    OS: '+vm.storage_profile.image_reference.offer+'\n')
+        win.addstr('    Type: '+vm.hardware_profile.vm_size+'\n')
+        for disk in vm.storage_profile.data_disks:
+            win.addstr('    Data disk: '+str(disk.disk_size_gb)+'GB\n')
         win.addstr('\n')
     win.addstr('\n')
+    if (len(procs) > 0):
+        win.addstr('Waiting for docker installs to finish\n')
+        for p in procs:
+            p.join()
 
 def drawDockerUpdate(win):
     win.clear()
@@ -451,15 +499,29 @@ def drawDockerUpdate(win):
             try:
                 for a in i['NetworkInterfaces']:
                     ip = a['Association']['PublicIp']
-                    dockerPS = runDockerPsAWS(ip, 'keys/testing6.pem')
+                    dockerPS = runDockerPsAWS(ip, AWS_KEY)
                     win.addstr(ip+'\n')
-                    win.addstr(str(dockerPS)+'\n\n')
+                    win.addstr(str(dockerPS).replace('\\r','').replace('\\n', '\n')+'\n\n')
                     win.refresh()
             except Exception as e:
                 win.addstr(str(e)+'\n')
                 pass
     #Azure
-
+    for vm in compute_client.virtual_machines.list_all():
+        states = compute_client.virtual_machines.instance_view('cis4010A2', vm.name, expand='instanceView').statuses
+        if (len(states) > 1):
+            state = states[1].display_status
+            if 'running' in state:
+                public_ip = network_client.public_ip_addresses.get('cis4010A2', vm.name+'-pub-ip').ip_address
+                try:
+                    win.addstr(vm.name +'@'+public_ip+'\n')
+                    win.refresh()
+                    dockerPS = runDockerPsAzure(public_ip, AZURE_KEY)
+                    win.addstr(str(dockerPS).replace('\\r','').replace('\\n', '\n')+'\n\n')
+                    win.refresh()
+                except Exception as e:
+                    win.addstr(str(e)+'\n')
+                    pass
 
     win.addstr("Press Enter to continue...\n")
     while 1:          
@@ -508,14 +570,14 @@ def main(win):
                     rebootVMs()
                     msg = "Rebooting instances"
                 except Exception as e:
-                    msg = "Failed to reboot instances "+e
+                    msg = "Failed to reboot instances "+str(e)
             elif key == 116:
                 try:
                     killVMs()
-                    time.sleep(1)
+                    win.timeout(1000)          
                     msg = "Terminating instances"
                 except Exception as e:
-                    msg = "Failed to terminate instances "+e
+                    msg = "Failed to terminate instances "+str(e)
             elif key > 0:
                 msg = ""
             
@@ -529,7 +591,6 @@ def main(win):
 #endregion
 
 if __name__ == "__main__":
-
     global ec2_c 
     ec2_c = boto3.client('ec2', region_name='us-east-1')
     global ec2
@@ -540,28 +601,3 @@ if __name__ == "__main__":
     network_client = get_client_from_auth_file(NetworkManagementClient, auth_path='credentials.json')
 
     curses.wrapper(main)
-
-    '''
-    instances = ec2.meta.client.describe_instance_status(IncludeAllInstances=True)['InstanceStatuses']
-    ids = []
-    for s in instances:
-        ids.append(s['InstanceId'])
-    descriptions = ec2.meta.client.describe_instances(InstanceIds=ids)['Reservations']
-    for r in descriptions:
-        for i in r['Instances']:
-            #print(i)
-            print(i['NetworkInterfaces'][0]['Association']['PublicIp'])
-            print(i['InstanceId'])
-            print(i['ImageId'])
-            print(i['InstanceType'])
-            for block in i['BlockDeviceMappings']:
-                print(ec2.Volume(block['Ebs']))
-                print(ec2.Volume(block['Ebs']['VolumeId']).size)
-
-    for r in descriptions:
-        for i in r['Instances']:
-            print(i['ImageId'])
-            print(i['InstanceId'])
-            print(i['InstanceType'])
-            print()
-    '''
